Generate Page Title with theme_generate_page_title()
To help those who are starting out, or for those that are lazy like me, this is a little something for you.
Let the CMS generate a page title for you.

If your theme has added support for the CMS to generate a title tag, then it will contain the new generated title from the new function.

+ Added theme_generate_page_title() function.
* Updated generated title tag to use new generated title.
* Updated test theme with example usage of this function.

The function currently accepts two optional parameters.
First parameter is a String for the landing page title for the root of your site/app.
Second parameter is a boolean for whether to append the name defined in config to the generated title.

I hope you enjoy this update as I certainly enjoyed developing it.